code review changes to theme and globalstyle
