+++ client/src/index.js
@@ -1,31 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import { Theme } from "./theme";
-import styled, { createGlobalStyle } from "styled-components";
+import App from "./App";
 
-const Title = styled.h1`
-  font-size: 1.5em;
-  color: ${Theme.color.mediumbrown};
-`;
+const app = <App />;
 
-const Logo = styled.img`
-  width: 150px;
-`;
-
-const Body = createGlobalStyle`
-  body {
-    background-color: ${Theme.color.lightgreen};
-  }
-`;
-
-const App = () => {
-  return (
-    <div>
-      <Body />
-      <Logo src={require("./img/icon.png")} alt="plant-logo"></Logo>
-      <Title>grow.</Title>
-    </div>
-  );
-};
-
-ReactDOM.render(React.createElement(App), document.getElementById("root"));
+ReactDOM.render(app, document.getElementById("root"));
