--- client/src/theme.js
@@ -1,16 +1,0 @@
-export const Theme = {
-  color: {
-    lightgreen: "#C9E4CA",
-    lightblue: "#3BADDD",
-    mediumgreen: "#6EB56E",
-    mediumbrown: "#443730",
-    darkblue: "#2E3F4C"
-  },
-  font: {},
-  mediaQuery: {
-    small: "600px",
-    medium: "768px",
-    large: "992px",
-    xlarge: "1220px"
-  }
-};
